better establishment dto / schema separation
