+++ back/src/adapters/primary/routers/DtoAndSchemas/v3/output/SearchImmersionResultPublicV3.schema.ts
@@ -2,10 +2,10 @@ import {
   absoluteUrlSchema,
   appellationCodeSchema,
   appellationLabelSchema,
+  businessCustomizedNameSchema,
   businessNameSchema,
   contactModeSchema,
   createPaginatedSchema,
-  customizedNameSchema,
   dateTimeIsoStringSchema,
   fitForDisabledWorkersSchema,
   geoPositionSchema,
@@ -34,7 +34,7 @@ const searchImmersionResultPublicV3SchemaCommon = z.object({
   siret: siretSchema,
   establishmentScore: z.number(),
   name: businessNameSchema,
-  customizedName: customizedNameSchema.optional(),
+  customizedName: businessCustomizedNameSchema.optional(),
   voluntaryToImmersion: z.boolean(),
   position: geoPositionSchema,
   address: z.object({
