+++ shared/src/admin/admin.schema.ts
@@ -2,7 +2,7 @@ import { z } from "zod";
 import { agencyIdSchema, agencyRoleSchema } from "../agency/agency.schema";
 import { conventionIdSchema } from "../convention/convention.schema";
 import { emailSchema } from "../email/email.schema";
-import { withEstablishmentBannishmentJustificationSchema } from "../establishment/bannedEstablishmentInformations";
+import { withEstablishmentBannishmentJustificationSchema } from "../establishment/bannedEstablishmentInformations.schema";
 import { siretSchema } from "../siret/siret.schema";
 import { userIdSchema } from "../user/user.schema";
 import { zStringMinLength1Max1024 } from "../utils/string.schema";
