+++ shared/src/convention/convention.dto.ts
@@ -17,8 +17,8 @@ import type {
   LegacyAssessmentDto,
 } from "../assessment/assessment.dto";
 import type { Email } from "../email/email.dto";
-import type { WithBannedEstablishmentInformations } from "../establishment/bannedEstablishmentInformations";
-import type { BusinessName } from "../establishment/businessName";
+import type { WithBannedEstablishmentInformations } from "../establishment/bannedEstablishmentInformations.dto";
+import type { BusinessName } from "../establishment/establishment.dto";
 import type { FtConnectIdentity } from "../federatedIdentities/federatedIdentity.dto";
 import type { DateFilter } from "../filters";
 import type {
