+++ shared/src/discussion/DiscussionBuilder.ts
@@ -9,7 +9,7 @@ import {
   type ImmersionObjective,
 } from "../convention/convention.dto";
 import type { Email } from "../email/email.dto";
-import type { BusinessName } from "../establishment/businessName";
+import type { BusinessName } from "../establishment/establishment.dto";
 import type { ContactMode } from "../formEstablishment/FormEstablishment.dto";
 import type {
   AppellationAndRomeDto,
