+++ shared/src/discussion/discussion.dto.ts
@@ -8,8 +8,8 @@ import type {
   LevelOfEducation,
 } from "../convention/convention.dto";
 import type { Email } from "../email/email.dto";
-import type { WithBannedEstablishmentInformations } from "../establishment/bannedEstablishmentInformations";
-import type { BusinessName } from "../establishment/businessName";
+import type { WithBannedEstablishmentInformations } from "../establishment/bannedEstablishmentInformations.dto";
+import type { BusinessName } from "../establishment/establishment.dto";
 import type { ContactMode } from "../formEstablishment/FormEstablishment.dto";
 import type { PhoneNumber } from "../phone/phone.dto";
 import type {
