+++ shared/src/establishment/bannedEstablishmentInformations.schema.ts
@@ -1,12 +1,8 @@
 import z from "zod";
-import type { WithSiretDto } from "../siret/siret";
 import { siretSchema } from "../siret/siret.schema";
 import { zStringMinLength1Max1024 } from "../utils/string.schema";
 import type { ZodSchemaWithInputMatchingOutput } from "../zodUtils";
-
-type WithEstablishmentBannishmentJustification = {
-  establishmentBannishmentJustification: string;
-};
+import type { WithBannedEstablishmentInformations } from "./bannedEstablishmentInformations.dto";
 
 const withEstablishmentBannishmentJustificationShape = {
   establishmentBannishmentJustification: zStringMinLength1Max1024,
@@ -16,12 +12,6 @@ export const withEstablishmentBannishmentJustificationSchema = z.object(
   withEstablishmentBannishmentJustificationShape,
 );
 
-export type WithBannedEstablishmentInformations =
-  | { isEstablishmentBanned: false }
-  | ({
-      isEstablishmentBanned: true;
-    } & WithEstablishmentBannishmentJustification);
-
 export const withBannedEstablishmentInformationSchema: ZodSchemaWithInputMatchingOutput<WithBannedEstablishmentInformations> =
   z.union([
     z.object({ isEstablishmentBanned: z.literal(false) }),
@@ -30,9 +20,6 @@ export const withBannedEstablishmentInformationSchema: ZodSchemaWithInputMatchin
       ...withEstablishmentBannishmentJustificationShape,
     }),
   ]);
-
-export type BanEstablishmentPayload = WithSiretDto &
-  WithEstablishmentBannishmentJustification;
 
 export const banEstablishmentPayloadSchema = z.object({
   siret: siretSchema,
