+++ shared/src/formEstablishment/FormEstablishment.dto.ts
@@ -3,11 +3,11 @@ import type { WithAcquisition } from "../acquisition.dto";
 import type { AddressAndPosition, LocationId } from "../address/address.dto";
 import type { ApiConsumerName } from "../apiConsumer/ApiConsumer";
 import type { Email } from "../email/email.dto";
-import type { WithBannedEstablishmentInformations } from "../establishment/bannedEstablishmentInformations";
+import type { WithBannedEstablishmentInformations } from "../establishment/bannedEstablishmentInformations.dto";
 import type {
   BusinessName,
   BusinessNameCustomized,
-} from "../establishment/businessName";
+} from "../establishment/establishment.dto";
 import type { GroupName } from "../group/group.dto";
 import type { NafDto } from "../naf/naf.dto";
 import type { PhoneNumber } from "../phone/phone.dto";
