+++ shared/src/formEstablishment/FormEstablishment.schema.ts
@@ -4,8 +4,8 @@ import { absoluteUrlSchema } from "../AbsoluteUrl";
 import { withAcquisitionSchema } from "../acquisition.dto";
 import { addressAndPositionSchema } from "../address/address.schema";
 import { emailSchema } from "../email/email.schema";
-import { withBannedEstablishmentInformationSchema } from "../establishment/bannedEstablishmentInformations";
-import { businessNameSchema } from "../establishment/businessName";
+import { withBannedEstablishmentInformationSchema } from "../establishment/bannedEstablishmentInformations.schema";
+import { businessNameSchema } from "../establishment/businessComponents.schema";
 import { nafSchema } from "../naf/naf.schema";
 import { phoneNumberSchema } from "../phone/phone.schema";
 import { withRemoteWorkModeSchema } from "../remoteWorkMode/remoteWorkMode.schema";
