+++ shared/src/index.ts
@@ -71,10 +71,10 @@ export * from "./environment";
 export * from "./errors/errors";
 export * from "./errors/ftConnectErrors";
 export * from "./errors/httpErrors";
-export * from "./establishment/bannedEstablishmentInformations";
-export * from "./establishment/businessAddress";
-export * from "./establishment/businessName";
-export * from "./establishment/establishment";
+export * from "./establishment/bannedEstablishmentInformations.dto";
+export * from "./establishment/bannedEstablishmentInformations.schema";
+export * from "./establishment/establishment.dto";
+export * from "./establishment/establishment.schema";
 export * from "./establishment/superEstablishment";
 export * from "./featureFlag/featureFlags.dto";
 export * from "./featureFlag/featureFlags.schema";
