+++ shared/src/siret/siret.ts
@@ -1,5 +1,7 @@
-import type { BusinessAddress } from "../establishment/businessAddress";
-import type { BusinessName } from "../establishment/businessName";
+import type {
+  BusinessAddress,
+  BusinessName,
+} from "../establishment/establishment.dto";
 import type { NafDto } from "../naf/naf.dto";
 import type { Flavor } from "../typeFlavors";
